+++ src/perfTest/groovy/com/monits/gradle/sca/performance/fixture/AbstractPerfTestFixture.groovy
@@ -29,11 +29,12 @@ import spock.util.environment.Jvm
 @CompileStatic
 abstract class AbstractPerfTestFixture extends Specification {
     // A sample of gradle versions to be considered in general testing - don't test anything below 2.8, it spams stdout
-    static final List<String> TESTED_GRADLE_VERSIONS = ['2.8', '2.14.1', '3.1'].asImmutable()
+    static final List<String> TESTED_GRADLE_VERSIONS = ['2.8', '2.14.1', GradleVersion.current().version].asImmutable()
     @SuppressWarnings(['DuplicateStringLiteral', 'UnnecessaryCast'])
     static final List<String> TESTED_GRADLE_VERSIONS_FOR_ANDROID = (['2.8', '2.14.1'] +
-        (Jvm.current.java8Compatible ? ['3.1'] : [] as List<String>)).takeRight(2).asImmutable()
-    static final String BASELINE_PLUGIN_VERSION = '"com.monits:static-code-analysis-plugin:2.2.1"'
+        (Jvm.current.java8Compatible ? [GradleVersion.current().version] : [] as List<String>))
+        .takeRight(2).asImmutable()
+    static final String BASELINE_PLUGIN_VERSION = '"com.monits:static-code-analysis-plugin:2.3.4"'
 
     static final int NUMBER_OF_CLASSES_TO_ANALYZE = 100
 
@@ -246,6 +247,6 @@ abstract class AbstractPerfTestFixture extends Specification {
     @SuppressWarnings('DuplicateNumberLiteral')
     String androidVersionForGradle(final String gradleVersion) {
         GradleVersion.version(gradleVersion) < GradleVersion.version('3.0') ?
-            DEFAULT_ANDROID_VERSION : '2.2.0'
+            DEFAULT_ANDROID_VERSION : '2.2.3'
     }
 }